Data of ileal microbiome and mucosal immune profiles in response to dietary UGAM in weaned goats
Astragalus membranaceus has been shown to improve growth performance and immunity in weaned ruminants. The current study investigated the supplementation of ultra-grinded A. membranaceus powder on microbiota and mucosal immunity in the ileum, and blood biochemistry of weaned goats. The present results highlighted that dietary supplementation of UGAM may have enhanced the ileum health of weaned goats by reducing inflammation factor expression and the relative abundance of pathogenic microbes, and enhancing total protein and albumin concentrations.